+++ hpc/nvidia/cuda/fakeCUDA/rand.hpp
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0
 /* Copyright (c) 2025 Rong Tao */
-#include "compat.hpp"
+#include "cuda_compat.h"
 
 #ifdef __cplusplus
 extern "C" {